+++ client/src/Components/Login/FacebookLogin.jsx
@@ -1,0 +1,15 @@
+import React from "react";
+
+const FacebookLogin = ({ width = 24, height = 24, fill = "#000" }) => {
+  return (
+    <svg>
+      <path>
+
+      </path>
+    </svg>
+  );
+};
+
+export default FacebookLogin;
+
+
