+++ package.json
@@ -5,6 +5,7 @@
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
     "mongoose": "^6.0.3",
+    "react-google-login": "^5.2.2",
     "redux-devtools-extension": "^2.13.9",
     "redux-thunk": "^2.3.0"
   },
